Renaming react component
