+++ client/components/user-home.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import SingleUser from './SingleUser'
-import OrderHistory from './OrderHistory'
+import OrderHistory from './SingleUserOrders'
 
 /**
  * COMPONENT
